Drop OMFSK Constant down to 1.41git add . That means 1.4 baud for every 2hz of bandwidth.

+++ options/options.go
@@ -8,7 +8,7 @@ type Options struct {
 
 	OMFSKConstant float64
 
-	NFFTPower int
+	NFFTPower float64
 
 	// Debug
 	NoiseLevel int
